Update SVGNativeImageData's size (#3802)
* Update SVGNativeImageData's size
* Updates the size of SVG image data using the owner frameBox when SVG without
  viewport is used as a background image.
* Use double type when calculating the aspect ratio of svg for more precision.
* Add internal test.
* Update test
* Update computePreferredWidth same as compute size

+++ src/core/layout/FrameBlockBoxInlineLayout.cpp
@@ -4495,7 +4495,8 @@ void FrameReplaced::computePreferredWidth(PreferredWidthContext& ctx)
             h = contentHeightAfterApplyingBoxSizing(h);
 
             if (hasAspectRatio) {
-                w = h * (intrinsicWidth / intrinsicHeight);
+                w = h *
+                    (intrinsicWidth.toDouble() / intrinsicHeight.toDouble());
             }
         }
